feat: extract methods to separate read model
