+++ src/lib/features/project/project-read-model-type.ts
@@ -1,0 +1,37 @@
+import type { ProjectMode } from '../../types';
+import type { IProjectQuery } from './project-store-type';
+
+export type ProjectForUi = {
+    id: string;
+    name: string;
+    description?: string;
+    health: number;
+    createdAt: Date;
+    mode: ProjectMode;
+    memberCount: number;
+    favorite: boolean;
+    archivedAt?: Date;
+    featureCount: number;
+    lastUsage: Date | null;
+    lastUpdated: Date | null;
+};
+
+export type ProjectForInsights = {
+    id: string;
+    health: number;
+    memberCount: number;
+    featureCount: number;
+    staleFeatureCount: number;
+    potentiallyStaleFeatureCount: number;
+    avgTimeToProduction: number;
+};
+
+export interface IProjectReadModel {
+    getForAdminUi(
+        query?: IProjectQuery,
+        userId?: number,
+    ): Promise<ProjectForUi[]>;
+    getProjectsForInsights(
+        query?: IProjectQuery,
+    ): Promise<ProjectForInsights[]>;
+}
